Refactoring: Multi Overhauled Configurations: All other py files import the config and installed sensors objects from operations_config.py (Load from disk & create config variable once) Created separate py files for config, installed sensors, file locations, trigger variances (converting from variances in config to outside) WIP Moved some functions to more appropriate py files Added IP, DB size, current datetime to operations_sensors Updated operations_upgrades and prepared for configuration changes coming soon
