+++ operations_modules/operations_commands.py
@@ -19,17 +19,12 @@
 import os
 from shutil import disk_usage
 
-from operations_modules import operations_config
-from operations_modules import operations_db
+from operations_modules.operations_config import current_config, version, installed_sensors, \
+    restart_sensor_services_command
+from operations_modules.operations_file_locations import last_updated_file_location
+from operations_modules.operations_db import write_to_sql_database, CreateOtherDataEntry
 from operations_modules import operations_logger
 from operations_modules import operations_sensors
-import sensor_modules.linux_os as linux_os
-import sensor_modules.raspberry_pi_system as raspberry_pi_system
-
-sensor_system = raspberry_pi_system.CreateRPSystem()
-sensor_os = linux_os.CreateLinuxSystem()
-
-bash_commands = operations_config.sensor_bash_commands
 
 
 def get_sensor_readings():
@@ -53,20 +48,19 @@ def get_sensor_readings():
 
 def get_system_information():
     """ Returns System Information needed for a Control Center 'System Report'. """
-    sensor_config = operations_config.get_installed_config()
     free_disk = disk_usage("/")[2]
 
     try:
-        str_sensor_data = str(sensor_os.get_hostname()) + \
-                          "," + str(sensor_os.get_ip()) + \
-                          "," + str(sensor_os.get_sys_datetime()) + \
-                          "," + str(sensor_os.get_uptime()) + \
-                          "," + str(operations_config.version) + \
-                          "," + str(round(sensor_system.cpu_temperature(), 2)) + \
+        str_sensor_data = operations_sensors.get_hostname() + \
+                          "," + operations_sensors.get_ip() + \
+                          "," + str(operations_sensors.get_system_datetime()) + \
+                          "," + str(operations_sensors.get_system_uptime()) + \
+                          "," + str(version) + \
+                          "," + str(round(float(operations_sensors.get_cpu_temperature()), 2)) + \
                           "," + str(round(free_disk / (2 ** 30), 2)) + \
-                          "," + str(sensor_os.get_sql_db_size()) + \
-                          "," + str(sensor_config.write_to_db) + \
-                          "," + str(sensor_config.enable_custom) + \
+                          "," + str(operations_sensors.get_db_size()) + \
+                          "," + str(current_config.write_to_db) + \
+                          "," + str(current_config.enable_custom) + \
                           "," + str(get_last_updated())
     except Exception as error:
         operations_logger.network_logger.error("Sensor reading failed - " + str(error))
@@ -77,18 +71,16 @@ def get_system_information():
 
 def get_config_information():
     """ Opens configuration file and returns it as a comma separated string. """
-    temp_config = operations_config.get_installed_config()
-    installed_sensors = operations_config.get_installed_sensors()
     str_installed_sensors = installed_sensors.get_installed_names_str()
 
     try:
-        tmp_str_config = str(temp_config.sleep_duration_interval) + \
-                         "," + str(temp_config.sleep_duration_trigger) + \
-                         "," + str(temp_config.write_to_db) + \
-                         "," + str(temp_config.enable_custom) + \
-                         "," + str(temp_config.acc_variance) + \
-                         "," + str(temp_config.mag_variance) + \
-                         "," + str(temp_config.gyro_variance) + \
+        tmp_str_config = str(current_config.sleep_duration_interval) + \
+                         "," + str(current_config.sleep_duration_trigger) + \
+                         "," + str(current_config.write_to_db) + \
+                         "," + str(current_config.enable_custom) + \
+                         "," + str(current_config.acc_variance) + \
+                         "," + str(current_config.mag_variance) + \
+                         "," + str(current_config.gyro_variance) + \
                          "," + str(str_installed_sensors)
 
     except Exception as error:
@@ -109,7 +101,7 @@ def get_sensor_log(log_file):
 def get_last_updated():
     """ Returns when the sensor programs were last updated and how. """
     try:
-        last_updated_file = open(operations_config.last_updated_file_location, "r")
+        last_updated_file = open(last_updated_file_location, "r")
         tmp_last_updated = last_updated_file.readlines()
         last_updated_file.close()
         last_updated = str(tmp_last_updated[0]) + str(tmp_last_updated[1])
@@ -122,11 +114,11 @@ def get_last_updated():
 
 def restart_services():
     """ Reloads systemd service files & restarts all sensor program services. """
-    os.system(operations_config.restart_sensor_services_command)
+    os.system(restart_sensor_services_command)
 
 
 def add_note_to_database(datetime_note):
-    sql_data = operations_db.CreateOtherDataEntry()
+    sql_data = CreateOtherDataEntry()
     datetime_note = datetime_note.strip()
     datetime_note = datetime_note.replace("'", '"')
     custom_datetime = "'" + datetime_note[:23] + "'"
@@ -139,4 +131,4 @@ def add_note_to_database(datetime_note):
                    sql_data.sql_query_values_start + sql_data.sensor_readings +
                    sql_data.sql_query_values_end)
 
-    operations_db.write_to_sql_database(sql_execute)
+    write_to_sql_database(sql_execute)
