+++ operations_modules/operations_db.py
@@ -20,14 +20,14 @@ import sqlite3
 
 from operations_modules import operations_sensors
 from operations_modules import operations_logger
-from operations_modules.operations_config import sensor_database_location
+import operations_modules.operations_file_locations as file_locations
 
 
 class CreateIntervalDatabaseData:
     """ Creates a object, holding required data for making a Interval SQL execute string. """
 
     def __init__(self):
-        self.database_location = sensor_database_location
+        self.database_location = file_locations.sensor_database_location
         self.sql_query_start = "INSERT OR IGNORE INTO IntervalData ("
         self.sql_query_values_start = ") VALUES ("
         self.sql_query_values_end = ")"
@@ -102,7 +102,7 @@ class CreateOtherDataEntry:
     """ Creates a object, holding required data for making a OtherData SQL execute string. """
 
     def __init__(self):
-        self.database_location = sensor_database_location
+        self.database_location = file_locations.sensor_database_location
         self.sql_query_start = "INSERT OR IGNORE INTO OtherData ("
         self.sql_query_values_start = ") VALUES ("
         self.sql_query_values_end = ")"
@@ -111,16 +111,83 @@ class CreateOtherDataEntry:
         self.sensor_readings = ""
 
 
+class CreateDatabaseVariables:
+    def __init__(self):
+        self.table_interval = "IntervalData"
+        self.table_trigger = "TriggerData"
+        self.table_other = "OtherData"
+
+        self.other_table_column_user_date_time = "UserDateTime"
+        self.other_table_column_notes = "Notes"
+
+        self.sensor_name = "SensorName"
+        self.ip = "IP"
+        self.sensor_uptime = "SensorUpTime"
+        self.system_temperature = "SystemTemp"
+        self.env_temperature = "EnvironmentTemp"
+        self.env_temperature_offset = "EnvTempOffset"
+        self.pressure = "Pressure"
+        self.humidity = "Humidity"
+        self.lumen = "Lumen"
+        self.red = "Red"
+        self.orange = "Orange"
+        self.yellow = "Yellow"
+        self.green = "Green"
+        self.blue = "Blue"
+        self.violet = "Violet"
+        self.acc_x = "Acc_X"
+        self.acc_y = "Acc_Y"
+        self.acc_z = "Acc_Z"
+        self.mag_x = "Mag_X"
+        self.mag_y = "Mag_Y"
+        self.mag_z = "Mag_Z"
+        self.gyro_x = "Gyro_X"
+        self.gyro_y = "Gyro_Y"
+        self.gyro_z = "Gyro_Z"
+
+    def get_sensor_columns_list(self):
+        sensor_sql_columns = [self.sensor_name,
+                              self.ip,
+                              self.sensor_uptime,
+                              self.system_temperature,
+                              self.env_temperature,
+                              self.env_temperature_offset,
+                              self.pressure,
+                              self.humidity,
+                              self.lumen,
+                              self.red,
+                              self.orange,
+                              self.yellow,
+                              self.green,
+                              self.blue,
+                              self.violet,
+                              self.acc_x,
+                              self.acc_y,
+                              self.acc_z,
+                              self.mag_x,
+                              self.mag_y,
+                              self.mag_z,
+                              self.gyro_x,
+                              self.gyro_y,
+                              self.gyro_z]
+        return sensor_sql_columns
+
+    def get_other_columns_list(self):
+        other_sql_columns = [self.other_table_column_user_date_time,
+                             self.other_table_column_notes]
+        return other_sql_columns
+
+
 def write_to_sql_database(sql_execute):
     """ Executes provided string with SQLite3.  Used to write sensor readings to the SQL Database. """
     operations_logger.primary_logger.debug("SQL String to execute: " + str(sql_execute))
 
     try:
-        db_connection = sqlite3.connect(sensor_database_location)
+        db_connection = sqlite3.connect(file_locations.sensor_database_location)
         db_cursor = db_connection.cursor()
         db_cursor.execute(sql_execute)
         db_connection.commit()
         db_connection.close()
-        operations_logger.primary_logger.debug("SQL Write to DataBase - OK - " + sensor_database_location)
+        operations_logger.primary_logger.debug("SQL Write to DataBase OK - " + file_locations.sensor_database_location)
     except Exception as error:
-        operations_logger.primary_logger.error("SQL Write to DataBase - Failed - " + str(error))
+        operations_logger.primary_logger.error("SQL Write to DataBase Failed - " + str(error))
