+++ record_to_db.py
@@ -21,17 +21,18 @@ from datetime import datetime
 from threading import Thread
 from time import sleep
 
-from operations_modules import operations_config
-from operations_modules import operations_db
+import operations_modules.operations_db as operations_db
 from operations_modules import operations_logger
 from operations_modules import operations_pre_checks
 from operations_modules import operations_sensors
+from operations_modules.operations_config import installed_sensors, current_config, trigger_variances, version, \
+    get_old_version, sense_hat_show_led_message, trigger_pairs
 
 # Ensure files, database & configurations are OK
 operations_pre_checks.check_missing_files()
 operations_pre_checks.check_database_structure()
 
-if operations_config.get_old_version() != operations_config.version:
+if get_old_version() != version:
     operations_logger.primary_logger.info("Checking files and configuration after upgrade")
     os.system("systemctl start SensorUpgradeChecks")
     # Sleep before loading anything due to needed updates
@@ -39,11 +40,9 @@ if operations_config.get_old_version() != operations_config.version:
     while True:
         sleep(10)
 
-installed_sensors = operations_config.get_installed_sensors()
-current_config = operations_config.get_installed_config()
-database_variables = operations_config.CreateDatabaseVariables()
-
 operations_logger.primary_logger.info("Sensor Recording to SQLite3 DB Started")
+
+database_variables = operations_db.CreateDatabaseVariables()
 
 
 def start_interval_recording():
@@ -58,7 +57,7 @@ def start_interval_recording():
 
             operations_db.write_to_sql_database(interval_sql_execute)
 
-            if installed_sensors.raspberry_pi_sense_hat and operations_config.sense_hat_show_led_message:
+            if installed_sensors.raspberry_pi_sense_hat and sense_hat_show_led_message:
                 operations_sensors.rp_sense_hat_sensor_access.display_led_message("SQL-Int-Rec")
         except Exception as error:
             operations_logger.primary_logger.error("Interval Failure: " + str(error))
@@ -66,7 +65,7 @@ def start_interval_recording():
         sleep(current_config.sleep_duration_interval)
 
 
-def _check_xyz_variances(trigger_data):
+def _check_against_variance(trigger_data):
     write_to_db = False
     pair_differences = []
 
@@ -74,44 +73,37 @@ def _check_xyz_variances(trigger_data):
     for reading in trigger_data.sql_readings1:
         readings_old = reading
         readings_new = trigger_data.sql_readings2[count]
-
-        count2 = 0
-        if count2 < 3:
-            try:
-                difference = abs(abs(float(readings_new[count2])) - abs(float(readings_old[count2])))
-                pair_differences.append(difference)
-            except Exception as error:
-                operations_logger.primary_logger.error("Bad readings in XYZ variance checks: " + str(error))
-                write_to_db = False
-
-            count += 1
+        try:
+            difference = abs(abs(float(readings_new)) - abs(float(readings_old)))
+            pair_differences.append(difference)
+        except Exception as error:
+            operations_logger.primary_logger.error("Bad readings in Single variance checks: " + str(error))
+            write_to_db = False
+        count += 1
 
     for difference in pair_differences:
         if difference >= trigger_data.variance:
-            operations_logger.primary_logger.debug("DIFFs Triggered")
             write_to_db = True
 
     if write_to_db:
         operations_logger.primary_logger.debug("Pair Differences: " + str(pair_differences))
-        sql_db_executes = trigger_data.get_xyz_sql_write_str()
+        sql_db_executes = trigger_data.get_sql_write_str()
 
         for execute in sql_db_executes:
             operations_db.write_to_sql_database(execute)
 
 
 def check_write_trigger_data():
-    operations_logger.primary_logger.debug("Check Triggers")
     sleep(current_config.sleep_duration_trigger)
 
-    threads = []
-    if installed_sensors.has_acc:
-        threads.append(Thread(target=_acc_check_thread))
+    sensor_selection = operations_db.CreateDatabaseVariables()
 
-    if installed_sensors.has_mag:
-        threads.append(Thread(target=_mag_check_thread))
+    threads = [Thread(target=_check_sensor, args=[sensor_selection.acc_x]),
+               Thread(target=_check_sensor, args=[sensor_selection.mag_x]),
+               Thread(target=_check_sensor, args=[sensor_selection.gyro_x])]
 
-    if installed_sensors.has_gyro:
-        threads.append(Thread(target=_gyro_check_thread))
+    # for sensor in sensor_selection.get_sensor_columns_list():
+    #     threads.append(Thread(target=_check_sensor, args=[sensor]))
 
     for thread in threads:
         thread.start()
@@ -120,55 +112,237 @@ def check_write_trigger_data():
         thread.join()
 
 
-def _acc_check_thread():
-    new_trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
-    new_trigger_data.variance = current_config.acc_variance
-    new_trigger_data.sql_columns_str += database_variables.get_acc_columns_str()
+def _check_sensor(sensor_type):
+    sensor_selection = operations_db.CreateDatabaseVariables()
+    trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
+
+    if sensor_type == sensor_selection.sensor_name:
+        trigger_data.variance = trigger_variances.sensor_name
+        trigger_data.sql_columns_str += database_variables.sensor_name
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_hostname())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_hostname())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.ip:
+        trigger_data.variance = trigger_variances.ip
+        trigger_data.sql_columns_str += database_variables.ip
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_ip())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_ip())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.sensor_uptime:
+        trigger_data.variance = trigger_variances.sensor_uptime
+        trigger_data.sql_columns_str += database_variables.sensor_uptime
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_system_uptime())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_system_uptime())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.system_temperature:
+        trigger_data.variance = trigger_variances.cpu_temperature
+        trigger_data.sql_columns_str += database_variables.system_temperature
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_cpu_temperature())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_cpu_temperature())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.env_temperature:
+        trigger_data.variance = trigger_variances.env_temperature
+        trigger_data.sql_columns_str += database_variables.env_temperature
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_sensor_temperature())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_sensor_temperature())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.pressure:
+        trigger_data.variance = trigger_variances.pressure
+        trigger_data.sql_columns_str += database_variables.pressure
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_pressure())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_pressure())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.humidity:
+        trigger_data.variance = trigger_variances.humidity
+        trigger_data.sql_columns_str += database_variables.humidity
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_humidity())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_humidity())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.lumen:
+        trigger_data.variance = trigger_variances.lumen
+        trigger_data.sql_columns_str += database_variables.lumen
+
+        pair_count = 0
+        while pair_count < trigger_pairs:
+            trigger_data.sql_readings1.append(operations_sensors.get_lumen())
+            trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
+            sleep(current_config.sleep_duration_trigger)
+
+            trigger_data.sql_readings2.append(operations_sensors.get_lumen())
+            trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+            pair_count += 1
+        # Need to create a new detection for this one
+        # _check_against_variance(trigger_data)
+    elif sensor_type == sensor_selection.red:
+        pass
+    elif sensor_type == sensor_selection.orange:
+        pass
+    elif sensor_type == sensor_selection.yellow:
+        pass
+    elif sensor_type == sensor_selection.green:
+        pass
+    elif sensor_type == sensor_selection.blue:
+        pass
+    elif sensor_type == sensor_selection.violet:
+        pass
+    elif sensor_type == sensor_selection.acc_x:
+        if installed_sensors.has_acc:
+            _check_xyz_thread(sensor_type)
+    elif sensor_type == sensor_selection.mag_x:
+        if installed_sensors.has_mag:
+            _check_xyz_thread(sensor_type)
+    elif sensor_type == sensor_selection.gyro_x:
+        if installed_sensors.has_gyro:
+            _check_xyz_thread(sensor_type)
+
+
+def _check_xyz_thread(sensor_type):
+    sensor_selection = operations_db.CreateDatabaseVariables()
+    x_trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
+    y_trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
+    z_trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
+    if sensor_type == sensor_selection.acc_x:
+        x_trigger_data.variance = current_config.acc_variance
+        x_trigger_data.sql_columns_str += database_variables.acc_x
+        y_trigger_data.variance = current_config.acc_variance
+        y_trigger_data.sql_columns_str += database_variables.acc_y
+        z_trigger_data.variance = current_config.acc_variance
+        z_trigger_data.sql_columns_str += database_variables.acc_z
+    elif sensor_type == sensor_selection.mag_x:
+        x_trigger_data.variance = current_config.mag_variance
+        x_trigger_data.sql_columns_str += database_variables.mag_x
+        y_trigger_data.variance = current_config.mag_variance
+        y_trigger_data.sql_columns_str += database_variables.mag_y
+        z_trigger_data.variance = current_config.mag_variance
+        z_trigger_data.sql_columns_str += database_variables.mag_z
+    elif sensor_type == sensor_selection.gyro_x:
+        x_trigger_data.variance = current_config.gyro_variance
+        x_trigger_data.sql_columns_str += database_variables.gyro_x
+        y_trigger_data.variance = current_config.gyro_variance
+        y_trigger_data.sql_columns_str += database_variables.gyro_y
+        z_trigger_data.variance = current_config.gyro_variance
+        z_trigger_data.sql_columns_str += database_variables.gyro_z
 
     pair_count = 0
-    while pair_count < operations_config.trigger_pairs:
-        new_trigger_data.sql_readings1.append(operations_sensors.get_accelerometer_xyz())
-        new_trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+    while pair_count < trigger_pairs:
+        if sensor_type == sensor_selection.acc_x:
+            xyz = operations_sensors.get_accelerometer_xyz()
+        elif sensor_type == sensor_selection.mag_x:
+            xyz = operations_sensors.get_magnetometer_xyz()
+        elif sensor_type == sensor_selection.gyro_x:
+            xyz = operations_sensors.get_gyroscope_xyz()
+        else:
+            xyz = [0, 0, 0]
+
+        x_trigger_data.sql_readings1.append(xyz[0])
+        x_trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+        y_trigger_data.sql_readings1.append(xyz[1])
+        y_trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+        z_trigger_data.sql_readings1.append(xyz[2])
+        z_trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
+
         sleep(current_config.sleep_duration_trigger)
-        new_trigger_data.sql_readings2.append(operations_sensors.get_accelerometer_xyz())
-        new_trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
+        if sensor_type == sensor_selection.acc_x:
+            xyz = operations_sensors.get_accelerometer_xyz()
+        elif sensor_type == sensor_selection.mag_x:
+            xyz = operations_sensors.get_magnetometer_xyz()
+        elif sensor_type == sensor_selection.gyro_x:
+            xyz = operations_sensors.get_gyroscope_xyz()
+        else:
+            xyz = [0, 0, 0]
+
+        x_trigger_data.sql_readings2.append(xyz[0])
+        x_trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+        y_trigger_data.sql_readings2.append(xyz[1])
+        y_trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+        z_trigger_data.sql_readings2.append(xyz[2])
+        z_trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
+
         pair_count += 1
 
-    _check_xyz_variances(new_trigger_data)
-
-
-def _mag_check_thread():
-    new_trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
-    new_trigger_data.variance = current_config.mag_variance
-    new_trigger_data.sql_columns_str += database_variables.get_mag_columns_str()
-
-    pair_count = 0
-    while pair_count < operations_config.trigger_pairs:
-        new_trigger_data.sql_readings1.append(operations_sensors.get_magnetometer_xyz())
-        new_trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
-        sleep(current_config.sleep_duration_trigger)
-        new_trigger_data.sql_readings2.append(operations_sensors.get_magnetometer_xyz())
-        new_trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
-        pair_count += 1
-
-    _check_xyz_variances(new_trigger_data)
-
-
-def _gyro_check_thread():
-    new_trigger_data = operations_db.CreateTriggerDatabaseData(installed_sensors)
-    new_trigger_data.variance = current_config.gyro_variance
-    new_trigger_data.sql_columns_str += database_variables.get_gyro_columns_str()
-
-    pair_count = 0
-    while pair_count < operations_config.trigger_pairs:
-        new_trigger_data.sql_readings1.append(operations_sensors.get_gyroscope_xyz())
-        new_trigger_data.sql_readings1_datetime.append(get_datetime_stamp())
-        sleep(current_config.sleep_duration_trigger)
-        new_trigger_data.sql_readings2.append(operations_sensors.get_gyroscope_xyz())
-        new_trigger_data.sql_readings2_datetime.append(get_datetime_stamp())
-        pair_count += 1
-
-    _check_xyz_variances(new_trigger_data)
+    _check_against_variance(x_trigger_data)
+    _check_against_variance(y_trigger_data)
+    _check_against_variance(z_trigger_data)
 
 
 def get_datetime_stamp():
@@ -176,13 +350,23 @@ def get_datetime_stamp():
 
 
 if current_config.write_to_db and installed_sensors.no_sensors is False:
-    # Start Interval Recording
-    interval_recording_thread = Thread(target=start_interval_recording)
-    interval_recording_thread.daemon = True
-    interval_recording_thread.start()
+    if current_config.enable_interval_recording:
+        # Start Interval Recording
+        interval_recording_thread = Thread(target=start_interval_recording)
+        interval_recording_thread.daemon = True
+        interval_recording_thread.start()
+    else:
+        operations_logger.primary_logger.warning("Interval Recording Disabled in Config")
+        while True:
+            sleep(600)
 
-    while True:
-        check_write_trigger_data()
+    if current_config.enable_trigger_recording:
+        while True:
+            check_write_trigger_data()
+    else:
+        operations_logger.primary_logger.warning("Trigger Recording Disabled in Config")
+        while True:
+            sleep(600)
 
 else:
     operations_logger.primary_logger.warning("Write to Database Disabled in Config or no sensors set")
