+++ sensor_modules/Temperature_Offsets.py
@@ -3,6 +3,7 @@ This module holds Environmental Temperature offsets based on hardware
 
 @author: OO-Dragon
 """
+from operations_modules.operations_config import current_config, installed_sensors
 
 
 class CreateRPZeroWTemperatureOffsets:
@@ -27,3 +28,29 @@ class CreateUnknownTemperatureOffsets:
         self.pimoroni_bme680 = 0.0
         self.pimoroni_enviro = 0.0
         self.rp_sense_hat = 0.0
+
+
+def get_sensor_temperature_offset():
+    """
+     Returns sensors Environmental temperature offset based on system board and sensor.
+     You can set an override in the main sensor configuration file.
+    """
+
+    if installed_sensors.raspberry_pi_3b_plus:
+        sensor_temp_offset = CreateRP3BPlusTemperatureOffsets()
+    elif installed_sensors.raspberry_pi_zero_w:
+        sensor_temp_offset = CreateRPZeroWTemperatureOffsets()
+    else:
+        # All offsets are 0.0 for unselected or unsupported system boards
+        sensor_temp_offset = CreateUnknownTemperatureOffsets()
+
+    if current_config.enable_custom_temp:
+        return current_config.custom_temperature_offset
+    elif installed_sensors.pimoroni_enviro:
+        return sensor_temp_offset.pimoroni_enviro
+    elif installed_sensors.pimoroni_bme680:
+        return sensor_temp_offset.pimoroni_bme680
+    elif installed_sensors.raspberry_pi_sense_hat:
+        return sensor_temp_offset.rp_sense_hat
+    else:
+        return 0.0
